Finish header content layout and styling
Placed items at their ideal position perfectly using Grid on and
inside the items of header element.

Used em units to scale the icons and images with respect to the text
font size thereby preventing the usage of calc().

Padding and gaps however, have been set in pixels so that they don't
scale in case the user is using a large base font size.

Applied min-width of 80px on search input so that it does not collapse
completely on narrow screens.

Generated columns of auto size for option buttons so that they won't
take all of the free space of container.

The changes mostly consist of basic stuff like alignment and basic
styling of items, hence there is no need to address all of them in
this commit.