Finish dashboard sidebar alignment and styling
Used invert(100%) filter on all icons so that they turn white from
black to go well with the blue background.

Applied padding reset on ul elements in order to get rid of unnecessary
initial browser styling.

Set cursor property to pointer value on all list items for better UX.

Gaps set in px instead of rems so that they don't over-scale in case
the user has set a larger font-size on their browser.

Rest of the styling involves application of Grid display and aligning
the elements as needed.