Fix projects panel icon and dashboard icon styling
Initially, the dashboard icon images were targeted using the img tag,
hence accessing all the img tags on the page. This led to the projects
panel images being invisible because of white background and their white
color as a result of filter invert being set on the img tag.

Fixed the issue by applying dashboard icon styling exclusively to the
dashboard icons by using a more specific selector.

Applied height and width to the project icon as needed.