Put trending item text content inside div
It was hectic to assign rows and columns separately to each text div
and image element.
To ease this by making the distribution automatic, all text div are
now nested inside the .id-container div.

With this, the grid space distribution is automatic as there are only
two elements: img and .id-container div.