Remove margin from header and use z-index in order to make box-shadow visible
Initially the margin was set in order to show the box-shadow of header
which was being covered by the main element.
However, the margin applied stood out a bit and looked a bit unrefined.

Therefore, applied a z-index value to the header element in order to
give it a higher preference in the element stacking order.
As a result, the box-shadow of header is now visible on top of the main
element because of the greater stack order of header due to z-index.